Fixes incoming...

+++ core/fileformats/ppmLoader.h
@@ -73,8 +73,8 @@ private:
     static string prefix1, prefix2;
 
     char* nextLine(FILE *fp, int sz, MetaData *metadata);
-    bool readHeader(FILE *fp, unsigned long int *h, unsigned long int *w, unsigned short int *maxval, uint8_t *type, MetaData* metadata);
-    bool writeHeader(FILE *fp, unsigned long int h, unsigned long int w, uint8_t type, uint64_t maxval, MetaData* metadata);
+    bool readHeader(FILE *fp, unsigned long int *h, unsigned long int *w, uint16_t *maxval, uint8_t *type, MetaData* metadata);
+    bool writeHeader(FILE *fp, unsigned long int h, unsigned long int w, uint8_t type, uint16_t maxval, MetaData* metadata);
 };
 
 } //namespace corecvs
